refactor: Reorganize seller services into Profile folder structure
- Move ISellerService → Services/Seller/Profile/ISellerProfileService
- Move SellerServiceSimple → Services/Seller/Profile/SellerProfileService
- Update namespace from Services.Seller to Services.Seller.Profile
- Update service registration in Program.cs and SellerController
- Rename class SellerServiceSimple → SellerProfileService for clarity

Benefits:
- Clear separation of seller profile management functionality
- Consistent with User/Profile structure pattern
- Better organization for future seller features (Products, Orders)
- More descriptive service names that indicate specific purpose
